LPS-62861 move to 6.0 since the global scope was added in 6.0
